Adding env & types file & fully factore
