+++ .env
@@ -1,0 +1,2 @@
+ REACT_APP_API_KEY= 19dc8c994b8ef838ba65a40c5ea44444
+
