--- src/context/apikey.ts
@@ -1,2 +1,0 @@
-const API_KEY = "19dc8c994b8ef838ba65a40c5ea44444";
-export default API_KEY;
